Initial Commit
Word2vec is a method to efficiently create word embeddings and has been
around since 2013. But in addition to its utility as a word-embedding
method, some of its concepts have been shown to be effective in creating
recommendation engines and making sense of sequential data even in
commercial, non-language tasks. Companies like Airbnb, Alibaba, Spotify,
and Anghami have all benefitted from carving out this brilliant piece of
machinery from the world of NLP and using it in production to empower a
new breed of recommendation engines.